[CRES-75] refactor: category api 분리 #73
