+++ src/hooks/queries/useGetCategories.ts
@@ -1,6 +1,6 @@
-import studyApi from '@apis/study/studyApi';
+import categoryApi from '@apis/category/categoryApi';
 import { useQuery } from '@tanstack/react-query';
 
 export const useGetCategories = () => {
-  return useQuery(['useGetCategories'], () => studyApi.getCategories());
+  return useQuery(['useGetCategories'], () => categoryApi.getCategories());
 };
